tests/wpt/web-platform-tests/old-tests/submission/Infraware/Forms/contents/Forms/input_width.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p>
    <h3>input_width</h3>
  </p>
  
  <hr>

  <div id="log"></div>

  <form method="post"
      enctype="application/x-www-form-urlencoded"
      action=""
      name="input_form">
  <p><input type='image' id='input_text'></p>
  </form>

  <script>
    
    var input_text = document.getElementById("input_text");
    input_text.width = 30;

    if (typeof(input_text.width) == "number") {
      test(function() {
        assert_equals(input_text.width, 30, "width attribute is not correct.");
      });
    } else {
      test(function() {
        assert_unreached("width attribute is not exist.");
      });
    }

  </script>

 </body>
</html>
